refs #101 Move event_slot_id selection to /events/register page
